Only delete a directory if it exists.
git-svn-id: https://vdt.cs.wisc.edu/svn/software/osg-test/trunk@14551 4e558342-562e-0410-864c-e07659590f8c

+++ osgtest/tests/test_78_voms.py
@@ -42,7 +42,8 @@ class TestStopVOMS(unittest.TestCase):
                      'VOMS server lock file still exists')
 
     def test_02_restore_vomses(self):
-        shutil.rmtree(core.config['voms.lsc-dir'])
+        if os.path.exists(core.config['voms.lsc-dir']):
+            shutil.rmtree(core.config['voms.lsc-dir'])
         files.restore('/etc/vomses')
 
     def test_03_remove_vo(self):
